TGBot/requirements.txt
telebot
python-dotenv
psycopg2-binary==2.8.5
jsonpickle
